requirements.txt
streamlit>=1.40.0
streamlit-drawable-canvas>=0.9.3
onnxruntime>=1.19.0
numpy>=1.26.0
pillow>=10.0.0
altair>=5.0.0
